Relax version requirements
Since `metalogistic` is a utility package, it will most likely be used in combination with many other packages.
Pinning dependency versions has the following issues:
- When users install the small metalog package into their environment, they suddenly
  see pip downloading new versions of numpy, scipy and matplotlib — ~100MB of downloads.
- If the users' application require a different version of numpy, scipy or matplotlib — tough luck.
- If any other package in the environment also pins dependency versions, then with almost
  100% certainty it cannot be installed alongside `metalogistic`.

The uses of numpy, scipy and matplotlib are extremely basic, have been
supported for over a decade, and will likely continue to be supported for
as long as those projects exist.
Thus I would argue that the best course is to remove the version requirements
completely, and let the user's other requirements dictate which versions they install.