Remove warnings when specifying bounds as -np.inf and np.inf
scipy.stats itself stores bounds as -np.inf and np.inf for unbounded functions,
so it seems unnecessary to discourage it.
One advantage is that then bounds can be given as `standard_bound + shift`,
and this works whether `standard_bound` is bounded or unbounded.
Also, if one were to eventually cythonize code, having bounds have a fixed
type of floats (rather than either float or None) is beneficial.