Implement auto-differentiable quantile functions.
The quantile function is extracted from `MetaLogistic` and stored as a
`MetalogSum` (list of `MetalogTerm` instances).
Each `MetalogTerm` can differentiate itself, returning 0 to 2 terms.
For unbounded metalogs, this is mostly sufficient to go to any order.
For bounded metalogs, we also need to differentiate through the log or
logit transformations – this needs a specific implementation for each order.

Requirements to differentiate up to order n:
- For unbounded dists: requires the derivative of ln(y/1-y) up to order n
- For bounded dists: requires in addition the derivative of the log/logit
  up to order n. There are three transforms; each must be implemented separately.

First attempt at auto-differentiating metalog